🌟feat: disruption enhancement (#109)
* 🌟feat: refactor disruption components for reusability

* 🌟feat: refactor more disruption components for reusability

* 🌟feat: add empty state for disruption

* 🌟feat: click category to filter in react-select for disruption

* 🌟feat: clean up debug message

# Conflicts:
#	src/locales/en/translation.json
#	src/locales/zh/translation.json
