+++ src/locales/en/translation.json
@@ -46,6 +46,10 @@
   "text.english": "English",
   "text.chinese": "中文",
   "text.map": "View map",
+  "text.not_found": "No results",
+  "search.sub_district": "District",
+  "search.location": "Flight/Location",
+  "search.placeholder": "Search for district/flight no./location",
   "dashboard.death_case": "Death",
   "dashboard.confirmed_case": "Confirmed",
   "dashboard.investigation_case": "Investigating",
