+++ src/locales/zh/translation.json
@@ -46,6 +46,10 @@
   "text.english": "English",
   "text.chinese": "中文",
   "text.map": "地圖",
+  "text.not_found": "沒有結果",
+  "search.sub_district": "地區",
+  "search.location": "航班/場所",
+  "search.placeholder": "搜尋地區/航班/場所",
   "dashboard.death_case": "死亡個案",
   "dashboard.confirmed_case": "確診個案",
   "dashboard.investigation_case": "住院檢查",
